refactor retro_analysis to a new module
