+++ run_lab.py
@@ -6,7 +6,7 @@ Then run `yarn start` or `python run_lab.py`
 import os
 # NOTE increase if needed. Pytorch thread overusage https://github.com/pytorch/pytorch/issues/975
 os.environ['OMP_NUM_THREADS'] = '1'
-from slm_lab.experiment import analysis
+from slm_lab.experiment import analysis, retro_analysis
 from slm_lab.experiment.control import Session, Trial, Experiment
 from slm_lab.experiment.monitor import InfoSpace
 from slm_lab.lib import logger, util
@@ -63,7 +63,7 @@ def run_old_mode(spec_file, spec_name, lab_mode):
         spec = spec_util.override_eval_spec(spec)
         Session(spec, info_space).run()
         util.clear_periodic_ckpt(prepath)  # cleanup after itself
-        analysis.analyze_eval_trial(spec, info_space, predir)
+        retro_analysis.analyze_eval_trial(spec, info_space, predir)
     else:
         raise ValueError('Unrecognizable lab_mode not of `enjoy, eval`')
